requirements.txt
aiobotocore==2.19.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.12.0
aiosignal==1.3.2
attrs==25.1.0
boto3==1.36.11
botocore==1.36.3
certifi==2025.1.31
cftime==1.6.4.post1
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2024.12.0
idna==3.10
jmespath==1.0.1
kiwisolver==1.4.8
matplotlib==3.10.0
multidict==6.1.0
netCDF4==1.7.2
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
pyparsing==3.2.1
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
s3fs==2024.12.0
s3transfer==0.11.2
scipy==1.15.1
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
wrapt==1.17.2
xarray==2025.1.2
yarl==1.18.3
